+++ Skynet/src/main/java/com/innowise/skynet/model/Factory.java
@@ -1,7 +1,7 @@
-package com.innowise.skynet.config.model;
+package com.innowise.skynet.model;
 
-import com.innowise.skynet.config.model.part.PartType;
-import com.innowise.skynet.config.model.part.RobotPart;
+import com.innowise.skynet.model.part.PartType;
+import com.innowise.skynet.model.part.RobotPart;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
