+++ Skynet/src/test/java/com/innowise/skynet/model/FactionTest.java
@@ -1,10 +1,8 @@
 package com.innowise.skynet.model;
 
 import com.innowise.skynet.config.SimulationConfig;
-import com.innowise.skynet.config.model.Faction;
-import com.innowise.skynet.config.model.Factory;
-import com.innowise.skynet.config.model.part.PartType;
-import com.innowise.skynet.config.model.part.RobotPart;
+import com.innowise.skynet.model.part.PartType;
+import com.innowise.skynet.model.part.RobotPart;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
