+++ Skynet/src/test/java/com/innowise/skynet/model/FactoryTest.java
@@ -1,7 +1,6 @@
 package com.innowise.skynet.model;
 
-import com.innowise.skynet.config.model.Factory;
-import com.innowise.skynet.config.model.part.RobotPart;
+import com.innowise.skynet.model.part.RobotPart;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
